🚚 Refactoriza estructura del proyecto: renombrado de `src` a `bib2graph` para mayor claridad y consistencia.
