+++ bib2graph/models.py
